Creating Dim_Sales

+++ models/staging/Sales/stg_erp_Sales_Order_Detail.sql
@@ -6,8 +6,8 @@ with
             , cast(orderqty as int) as OrderProductQuantity
             , cast(productid as int) as ProductId
             , cast(specialofferid  as int) as SpecialOfferId
-            , cast(unitprice as float) as ProductUnitPrice
-            , cast(unitpricediscount as float) as ProductUnitPriceDiscount
+            , cast(unitprice as float64) as ProductUnitPrice
+            , cast(unitpricediscount as float64) as ProductUnitPriceDiscount
         from {{ source('sales', 'salesorderdetail') }}
     )
